8 need fixed simulation time

+++ Data.cpp
@@ -3,7 +3,7 @@
 
 int SizeWindowX=800;
 int SizeWindowY=600;
-float mCurrentTick=0.0;
+int mCurrentTick=0;
 int myTick = 0;
 Camera cam1(vec3(0,1,50));
 vector<vec3> mTeapotColors;
